LPS-189884 Remove service from ReleaseManagerImpl, create ReleaseManagerHelper, apply to usages and fix integration test
