+++ modules/apps/portal/portal-upgrade-impl/src/main/java/com/liferay/portal/upgrade/internal/release/helper/ReleaseManagerHelper.java
@@ -1,0 +1,150 @@
+/**
+ * SPDX-FileCopyrightText: (c) 2023 Liferay, Inc. https://liferay.com
+ * SPDX-License-Identifier: LGPL-2.1-or-later OR LicenseRef-Liferay-DXP-EULA-2.0.0-2023-06
+ */
+
+package com.liferay.portal.upgrade.internal.release.helper;
+
+import com.liferay.osgi.service.tracker.collections.map.PropertyServiceReferenceComparator;
+import com.liferay.osgi.service.tracker.collections.map.PropertyServiceReferenceMapper;
+import com.liferay.osgi.service.tracker.collections.map.ServiceTrackerMap;
+import com.liferay.osgi.service.tracker.collections.map.ServiceTrackerMapFactory;
+import com.liferay.portal.kernel.model.Release;
+import com.liferay.portal.kernel.service.ReleaseLocalService;
+import com.liferay.portal.kernel.upgrade.UpgradeStep;
+import com.liferay.portal.kernel.util.GetterUtil;
+import com.liferay.portal.kernel.util.Validator;
+import com.liferay.portal.upgrade.internal.graph.ReleaseGraphManager;
+import com.liferay.portal.upgrade.internal.registry.UpgradeInfo;
+
+import java.util.Collections;
+import java.util.HashSet;
+import java.util.List;
+import java.util.Set;
+
+import org.osgi.framework.BundleContext;
+import org.osgi.framework.ServiceReference;
+import org.osgi.service.component.annotations.Activate;
+import org.osgi.service.component.annotations.Component;
+import org.osgi.service.component.annotations.Deactivate;
+import org.osgi.service.component.annotations.Reference;
+import org.osgi.util.tracker.ServiceTrackerCustomizer;
+
+/**
+ * @author Regisson Aguiar
+ */
+@Component(service = ReleaseManagerHelper.class)
+public class ReleaseManagerHelper {
+
+	public Set<String> getBundleSymbolicNames() {
+		return new HashSet<>(_serviceTrackerMap.keySet());
+	}
+
+	public String getSchemaVersionString(String bundleSymbolicName) {
+		Release release = _releaseLocalService.fetchRelease(bundleSymbolicName);
+
+		if ((release != null) &&
+			Validator.isNotNull(release.getSchemaVersion())) {
+
+			return release.getSchemaVersion();
+		}
+
+		return "0.0.0";
+	}
+
+	public Set<String> getUpgradableBundleSymbolicNames() {
+		Set<String> upgradableBundleSymbolicNames = new HashSet<>();
+
+		for (String bundleSymbolicName : getBundleSymbolicNames()) {
+			if (isUpgradable(bundleSymbolicName)) {
+				upgradableBundleSymbolicNames.add(bundleSymbolicName);
+			}
+		}
+
+		return upgradableBundleSymbolicNames;
+	}
+
+	public List<UpgradeInfo> getUpgradeInfos(String bundleSymbolicName) {
+		return _serviceTrackerMap.getService(bundleSymbolicName);
+	}
+
+	public boolean isUpgradable(String bundleSymbolicName) {
+		ReleaseGraphManager releaseGraphManager = new ReleaseGraphManager(
+			getUpgradeInfos(bundleSymbolicName));
+
+		List<List<UpgradeInfo>> upgradeInfosList =
+			releaseGraphManager.getUpgradeInfosList(
+				getSchemaVersionString(bundleSymbolicName));
+
+		if (upgradeInfosList.size() == 1) {
+			return true;
+		}
+
+		return false;
+	}
+
+	@Activate
+	protected void activate(BundleContext bundleContext) {
+		_serviceTrackerMap = ServiceTrackerMapFactory.openMultiValueMap(
+			bundleContext, UpgradeStep.class, null,
+			new PropertyServiceReferenceMapper<>(
+				"upgrade.bundle.symbolic.name"),
+			new UpgradeServiceTrackerCustomizer(bundleContext),
+			Collections.reverseOrder(
+				new PropertyServiceReferenceComparator<>(
+					"upgrade.from.schema.version")));
+	}
+
+	@Deactivate
+	protected void deactivate() {
+		_serviceTrackerMap.close();
+	}
+
+	@Reference
+	private ReleaseLocalService _releaseLocalService;
+
+	private ServiceTrackerMap<String, List<UpgradeInfo>> _serviceTrackerMap;
+
+	private static class UpgradeServiceTrackerCustomizer
+		implements ServiceTrackerCustomizer<UpgradeStep, UpgradeInfo> {
+
+		public UpgradeServiceTrackerCustomizer(BundleContext bundleContext) {
+			_bundleContext = bundleContext;
+		}
+
+		@Override
+		public UpgradeInfo addingService(
+			ServiceReference<UpgradeStep> serviceReference) {
+
+			String fromSchemaVersionString =
+				(String)serviceReference.getProperty(
+					"upgrade.from.schema.version");
+			String toSchemaVersionString = (String)serviceReference.getProperty(
+				"upgrade.to.schema.version");
+			int buildNumber = GetterUtil.getInteger(
+				serviceReference.getProperty("build.number"));
+
+			return new UpgradeInfo(
+				fromSchemaVersionString, toSchemaVersionString, buildNumber,
+				_bundleContext.getService(serviceReference));
+		}
+
+		@Override
+		public void modifiedService(
+			ServiceReference<UpgradeStep> serviceReference,
+			UpgradeInfo upgradeInfo) {
+		}
+
+		@Override
+		public void removedService(
+			ServiceReference<UpgradeStep> serviceReference,
+			UpgradeInfo upgradeInfo) {
+
+			_bundleContext.ungetService(serviceReference);
+		}
+
+		private final BundleContext _bundleContext;
+
+	}
+
+}
